# docs/swagger.yaml
basePath: /
definitions:
...
  httptransport.DeleteMenuRequest:
    properties:
      id:
        type: string
      version:
        type: integer
    required:
    - id
    - version
    type: object
...
  httptransport.MenuIDRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  httptransport.MenuMutationRequest:
    properties:
      component:
        type: string
      external_url:
        type: string
      icon:
        type: string
      id:
        type: string
      menu_key:
        type: string
      menu_type:
        type: string
      metadata:
        type: object
      name:
        type: string
      parent_id:
        type: string
      permission_id:
        type: string
      route_path:
        type: string
      sort_order:
        type: integer
      status:
        type: string
      version:
        type: integer
      visible:
        type: boolean
    required:
    - menu_type
    - name
    - status
    type: object
  httptransport.MenuTreeRequest:
    properties:
      created_at_from:
        type: string
      created_at_to:
        type: string
      ids:
        items:
          type: string
        type: array
      keyword:
        type: string
      statuses:
        items:
          type: string
        type: array
      types:
        items:
          type: string
        type: array
    type: object
...
  httptransport.Response:
    properties:
      body: {}
      code:
        type: integer
      message:
        type: string
      request_id:
        type: string
    type: object
...
  menu.Node:
    properties:
      children:
        items:
          $ref: '#/definitions/menu.Node'
        type: array
      component:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      external_url:
        type: string
      icon:
        type: string
      id:
        type: string
      menu_key:
        type: string
      menu_type:
        type: string
      metadata:
        type: object
      name:
        type: string
      parent_id:
        type: string
      permission_id:
        type: string
      route_path:
        type: string
      sort_order:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      version:
        type: integer
      visible:
        type: boolean
    type: object
  menu.Record:
    properties:
      component:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      external_url:
        type: string
      icon:
        type: string
      id:
        type: string
      menu_key:
        type: string
      menu_type:
        type: string
      metadata:
        type: object
      name:
        type: string
      parent_id:
        type: string
      permission_id:
        type: string
      route_path:
        type: string
      sort_order:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      version:
        type: integer
      visible:
        type: boolean
    type: object
...
info:
  contact: {}
  description: 'Production-oriented Go Web API scaffold. Application error codes:
    0 success; 10000-19999 common/input; 20000-29999 authentication/authorization;
    30000-39999 business; 50000-59999 infrastructure.'
  title: Go API Template
  version: "1.0"
paths:
...
  /api/v1/me/menus:
    post:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/menu.Node'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Return menus visible in the current tenant context
      tags:
      - menus
  /api/v1/menus/create:
    post:
      parameters:
      - description: Tree filters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.MenuTreeRequest'
      - description: Menu
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.MenuMutationRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/menu.Record'
              type: object
      security:
      - Bearer: []
      summary: Create a platform menu with a stable UUID v5 derived from menu_key
      tags:
      - menus
  /api/v1/menus/delete:
    post:
      parameters:
      - description: Menu and version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.DeleteMenuRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Delete a leaf platform menu
      tags:
      - menus
  /api/v1/menus/get:
    post:
      parameters:
      - description: Menu
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.MenuIDRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/menu.Record'
              type: object
      security:
      - Bearer: []
      summary: Get a platform menu
      tags:
      - menus
  /api/v1/menus/tree:
    post:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/menu.Node'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Return the complete platform menu tree
      tags:
      - menus
  /api/v1/menus/update:
    post:
      parameters:
      - description: Menu and version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.MenuMutationRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/menu.Record'
              type: object
      security:
      - Bearer: []
      summary: Update or move a platform menu
      tags:
      - menus
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Enter "Bearer {JWT}".
    in: header
    name: Authorization
    type: apiKey
  PSK:
    description: Enter "PSK {shared-key}" for routes configured with PSK authentication.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"